requirements.txt
requests
Pillow
telnetlib3
pyspellchecker